@@ -45,37 +45,60 @@
         "prepublishOnly": "npm run package",
         "preview": "vite preview",
         "test": "vitest run --coverage",
+        "test:all": "npm run test && npm run test:e2e",
+        "test:e2e": "playwright test",
+        "test:e2e:debug": "playwright test --debug",
+        "test:e2e:report": "playwright show-report",
+        "test:e2e:ui": "playwright test --ui",
         "test:only": "vitest run",
+        "test:unit": "vitest",
+        "test:unit:ui": "vitest --ui",
         "test:watch": "vitest"
     },
     "dependencies": {
         "@humanspeak/svelte-subscribe": "^5.0.0"
     },
     "devDependencies": {
-        "@sveltejs/adapter-auto": "^3.0.0",
-        "@sveltejs/kit": "^2.0.0",
-        "@sveltejs/package": "^2.0.0",
-        "@sveltejs/vite-plugin-svelte": "^3.0.0",
-        "@testing-library/jest-dom": "^6.1.6",
-        "@testing-library/svelte": "^4.0.5",
-        "@testing-library/user-event": "^14.5.2",
-        "@types/eslint": "8.56.0",
-        "@typescript-eslint/eslint-plugin": "^6.0.0",
-        "@typescript-eslint/parser": "^6.0.0",
-        "@vitest/coverage-v8": "^1.1.1",
-        "eslint": "^8.56.0",
-        "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-svelte": "^2.35.1",
-        "jsdom": "^23.0.1",
-        "prettier": "^3.1.1",
-        "prettier-plugin-svelte": "^3.1.2",
-        "publint": "^0.1.9",
-        "svelte": "^4.2.7",
-        "svelte-check": "^3.6.0",
-        "tslib": "^2.4.1",
-        "type-fest": "^4.9.0",
-        "typescript": "^5.0.0",
-        "vite": "^5.0.3",
-        "vitest": "^1.0.0"
+        "@eslint/compat": "^1.2.7",
+        "@eslint/js": "^9.21.0",
+        "@sveltejs/adapter-auto": "^4.0.0",
+        "@sveltejs/kit": "^2.19.0",
+        "@sveltejs/package": "^2.3.10",
+        "@sveltejs/vite-plugin-svelte": "^5.0.3",
+        "@testing-library/jest-dom": "^6.6.3",
+        "@testing-library/svelte": "^5.2.7",
+        "@testing-library/user-event": "^14.6.1",
+        "@types/eslint": "9.6.1",
+        "@types/node": "^22.13.9",
+        "@typescript-eslint/eslint-plugin": "^8.26.0",
+        "@typescript-eslint/parser": "^8.26.0",
+        "@vitest/coverage-v8": "^3.0.8",
+        "eslint": "^9.21.0",
+        "eslint-config-prettier": "^10.1.1",
+        "eslint-plugin-import": "^2.31.0",
+        "eslint-plugin-svelte": "^3.0.3",
+        "eslint-plugin-unused-imports": "^4.1.4",
+        "jsdom": "^26.0.0",
+        "prettier": "^3.5.3",
+        "prettier-plugin-organize-imports": "^4.1.0",
+        "prettier-plugin-svelte": "^3.3.3",
+        "prettier-plugin-tailwindcss": "^0.6.11",
+        "publint": "^0.3.8",
+        "svelte": "^5.22.6",
+        "svelte-check": "^4.1.5",
+        "tslib": "^2.8.1",
+        "type-fest": "^4.37.0",
+        "typescript": "^5.8.2",
+        "typescript-eslint": "^8.26.0",
+        "vite": "^6.2.1",
+        "vitest": "^3.0.8"
+    },
+    "peerDependencies": {
+        "svelte": "^5.0.0"
+    },
+    "overrides": {
+        "@sveltejs/kit": {
+            "cookie": "^0.7.0"
+        }
     }
 }
