@@ -1,21 +1,23 @@
 <script lang="ts">
-    import type { ComponentProps, SvelteComponent } from 'svelte'
+    import type { Component, ComponentProps } from 'svelte'
     import type { ComponentRenderConfig } from './createRender.js'
     import Render from './Render.svelte'
 
-    type TComponent = $$Generic<SvelteComponent>
+    // trunk-ignore(eslint/@typescript-eslint/no-explicit-any)
+    // trunk-ignore(eslint/no-undef)
+    type TComponent = $$Generic<Component<any>>
+    type Props = {
+        instance: TComponent | undefined
+        config: Omit<ComponentRenderConfig<TComponent>, 'props'>
+        props: ComponentProps<TComponent> | undefined
+    }
 
-    export let instance: TComponent | undefined = undefined
-    export let config: Omit<ComponentRenderConfig<TComponent>, 'props'>
-    export let props: ComponentProps<TComponent> | undefined = undefined
+    // trunk-ignore(eslint/prefer-const)
+    let { instance = $bindable(undefined), config, props }: Props = $props()
 </script>
 
-{#if config.children.length === 0}
-    <svelte:component this={config.component} bind:this={instance} {...props ?? {}} />
-{:else}
-    <svelte:component this={config.component} bind:this={instance} {...props ?? {}}>
-        {#each config.children as child}
-            <Render of={child} />
-        {/each}
-    </svelte:component>
-{/if}
+<config.component {...props} bind:this={instance}>
+    {#each config.children as child, i (i)}
+        <Render of={child} />
+    {/each}
+</config.component>
