@@ -2,6 +2,6 @@
     export let times: number
 </script>
 
-{#each { length: times } as _}
+{#each { length: times } as _, i (i)}
     🚀
 {/each}
