@@ -1,7 +1,10 @@
 <script lang="ts">
     import { Render, type RenderConfig } from '$lib/index.js'
+    import type { Component } from 'svelte'
 
-    type TComponent = $$Generic<SvelteComponent>
+    // trunk-ignore(eslint/@typescript-eslint/no-explicit-any)
+    // trunk-ignore(eslint/no-undef)
+    type TComponent = $$Generic<Component<any>>
 
     export let config: RenderConfig<TComponent>
 </script>
